requirements.txt
streamlit>=1.32.0
fastapi>=0.109.0
starlette>=0.36.0
pydantic==2.9.1
protobuf==4.25.3
pillow==10.3.0
rich==13.7.1
httpx>=0.27.0
psycopg2-binary>=2.9.9
tenacity>=8.2.3

